refactor: move extract_renderer_version to src.data.vst.core
The extractor reads VST3 plugin bundle metadata — that's a VST utility,
not a spec-schema concern. Move it next to the other VST helpers
(`load_plugin`, `load_preset`, `render_params`) in `src/data/vst/core.py`
and update the worker-side caller in `pipeline.entrypoints.generate_dataset`
to import from the new location.

`SURGE_XT_RENDERER_VERSION` stays in `pipeline.schemas.spec` because it
is a spec-construction constant (consumed by `materialize_spec`); only
the extractor moves. Tests follow the source: `TestExtractRendererVersion`
moves from `tests/pipeline/test_schemas/test_spec.py` to a new
`tests/data/vst/test_core.py` (matching the existing
`tests/data/vst/{test_generate_vst_dataset,test_preset_*}.py` layout).

No behavior change. The function signature and error contract are
identical; tests are byte-for-byte the same as their previous
location, just imported from the new path.

Refs #534
